move ExtendedValue to util module
